bugfixes for voyage

+++ code/game/turfs/floor_types.dm
@@ -283,19 +283,23 @@
 	..()
 	for(var/obj/covers/C in src)
 		return
+	for(var/obj/structure/multiz/ladder/S in src)
+		return
 	if (!istype(A, /mob) && !istype(A, /obj/structure/fishing_cage) && !istype(A, /obj/covers) && !istype(A, /obj/structure/barricade) && !!istype(A, /obj/effect/sailing_effect))
 		spawn(60)
-			if (A.loc == src)
+			if (A && A.loc == src)
 				qdel(A)
 	if(istype(A, /mob/living))
 		var/mob/living/ML = A
-		if (ML.stat == DEAD)
+		if (ML && ML.stat == DEAD)
 			spawn(60)
-				if (A.loc == src)
+				if (A && A.loc == src)
 					qdel(A)
 /turf/floor/beach/water/deep/Entered(atom/movable/A)
 	..()
 	for(var/obj/covers/C in src)
+		return
+	for(var/obj/structure/multiz/ladder/S in src)
 		return
 	if (!istype(A, /mob) && !istype(A, /obj/structure/fishing_cage) && !istype(A, /obj/covers) && !istype(A, /obj/structure/barricade) && !istype(A, /obj/effect/sailing_effect))
 		spawn(60)
@@ -303,9 +307,9 @@
 				qdel(A)
 	if(istype(A, /mob/living))
 		var/mob/living/ML = A
-		if (ML.stat == DEAD)
+		if (ML && ML.stat == DEAD)
 			spawn(60)
-				if (A.loc == src)
+				if (A && A.loc == src)
 					qdel(A)
 /turf/floor/beach/water/deep/jungle
 	name = "deep jungle river"
@@ -346,6 +350,8 @@
 
 /turf/proc/iscovered()
 	for(var/obj/covers/C in src)
+		return TRUE
+	for(var/obj/structure/multiz/ladder/ST in src)
 		return TRUE
 	for (var/obj/structure/vehicleparts/frame/ship/S in src)
 		return TRUE
